moved code gen functions into specializer
